backend/requirements.txt
aniso8601==9.0.1
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.1
cycler==0.12.1
debugpy==1.7.0
decorator==5.1.1
dnspython==2.4.2
email-validator==2.0.0.post2
executing==1.2.0
Flask==2.3.1
flask-babel==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-restx==1.1.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.3
Flask-Table==0.5.0
Flask-WTF==1.1.1
fonttools==4.43.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.7.0
PyMySQL==1.1.0
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3
pyzmq==25.1.1
scikit-learn==1.3.0
scipy==1.11.2
simple-websocket==1.0.0
six==1.16.0
sklearn==0.0.post9
SQLAlchemy==2.0.12
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
wcwidth==0.2.6
Werkzeug==2.3.2
wsproto==1.2.0
WTForms==3.0.1
